# Optimize Spy::Subrouting#get_spy_id